+++ internal/controller/config/config.go
@@ -6,8 +6,8 @@ import (
 	"sync"
 
 	"github.com/fsnotify/fsnotify"
+	apiv1 "github.com/numaproj-labs/numaplane/api/v1alpha1"
 	"github.com/spf13/viper"
-	corev1 "k8s.io/api/core/v1"
 )
 
 type ConfigManager struct {
@@ -36,33 +36,7 @@ type GlobalConfig struct {
 	ClusterName     string `json:"clusterName"`
 	TimeIntervalSec uint   `json:"timeIntervalSec"`
 	// RepoCredentials maps each Git Repository Path prefix to the corresponding credentials that are needed for it
-	RepoCredentials []RepoCredential `json:"repoCredentials"`
-}
-
-type RepoCredential struct {
-	URL            string          `json:"url"`
-	HTTPCredential *HTTPCredential `json:"httpCredential"`
-	SSHCredential  *SSHCredential  `json:"sshCredential"`
-	TLS            *TLS            `json:"tls"`
-}
-
-type HTTPCredential struct {
-	Username string            `json:"username"`
-	Password SecretKeySelector `json:"password"`
-}
-
-type SSHCredential struct {
-	SSHKey SecretKeySelector `json:"SSHKey" yaml:"SSHKey" `
-}
-
-type TLS struct {
-	InsecureSkipVerify bool `json:"insecureSkipVerify"`
-}
-
-type SecretKeySelector struct {
-	corev1.LocalObjectReference `mapstructure:",squash"` // for viper to correctly parse the config
-	Key                         string                   `json:"key" `
-	Optional                    *bool                    `json:"optional,omitempty" `
+	RepoCredentials []apiv1.RepoCredential `json:"repoCredentials"`
 }
 
 func (cm *ConfigManager) GetConfig() (GlobalConfig, error) {
